add more tests for pricerules
