--- pricerule/pricerule_voucher_cumulation_employees_test.go
@@ -1,1 +1,0 @@
-package pricerule
